* Source/cairo/CairoGState.m           abs -> fabs         * Source/x11/XGGLContext.m           format string fix         * Headers/x11/XGOpenGL.h           glxminorversion is int, not long int
Looks good, Ivan Vucica



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@38554 72102866-910b-0410-8b05-ffd578937521

+++ Source/x11/XGGLContext.m
@@ -506,7 +506,7 @@ static XGGLContext *currentGLContext;
   saved_ignores_backing = [view _ignoresBacking];
   [view _setIgnoresBacking: YES];
 
-  NSDebugMLLog(@"GLX", @"glx_window : %u", glx_drawable);
+  NSDebugMLLog(@"GLX", @"glx_window : %lu", (unsigned long) glx_drawable);
 }
 
 - (void)update
